Make the code more compact, move LastSpoken into getlinks
